+++ modules/Render/RenderVulkan/src/Resources/VkMaterialDescriptor.c
@@ -179,14 +179,14 @@ dsVkMaterialDescriptor* dsVkMaterialDescriptor_create(dsRenderer* renderer, dsAl
 }
 
 bool dsVkMaterialDescriptor_shouldCheckPointers(const dsVkMaterialDescriptor* descriptor,
-	const dsVkSamplerList* samplers, const void* refObject, uint32_t pointerVersion)
+	const dsVkSamplerList* samplers, const void* refObject, uint64_t pointerVersion)
 {
 	return descriptor->samplers != samplers || descriptor->refObject != refObject ||
 		descriptor->pointerVersion != pointerVersion;
 }
 
 bool dsVkMaterialDescriptor_shouldCheckOffsets(
-	const dsVkMaterialDescriptor* descriptor, uint32_t offsetVersion)
+	const dsVkMaterialDescriptor* descriptor, uint64_t offsetVersion)
 {
 	return descriptor->offsetVersion != offsetVersion;
 }
@@ -204,8 +204,8 @@ bool dsVkMaterialDescriptor_isUpToDate(
 }
 
 void dsVkMaterialDescriptor_updateEarlyChecks(dsVkMaterialDescriptor* descriptor,
-	const dsVkSamplerList* samplers, const void* refObject, uint32_t pointerVersion,
-	uint32_t offsetVersion)
+	const dsVkSamplerList* samplers, const void* refObject, uint64_t pointerVersion,
+	uint64_t offsetVersion)
 {
 	descriptor->samplers = samplers;
 	descriptor->refObject = refObject;
@@ -215,7 +215,7 @@ void dsVkMaterialDescriptor_updateEarlyChecks(dsVkMaterialDescriptor* descriptor
 
 void dsVkMaterialDescriptor_update(dsVkMaterialDescriptor* descriptor, const dsShader* shader,
 	dsVkBindingMemory* bindingMemory, const dsVkSamplerList* samplers, const void* refObject,
-	uint32_t pointerVersion, uint32_t offsetVersion)
+	uint64_t pointerVersion, uint64_t offsetVersion)
 {
 	DS_UNUSED(shader);
 	DS_ASSERT(shader->materialDesc == descriptor->materialDesc);
